overlay: Disable immersive mode confirmation
The confirmation dialogue is broken in several apps. Displaying the
confirmation dialogue leads to a focus change on which the app requests
immersive mode again leading to a loop of focus changes between the app
and the dialogue. Set the default value to confirmed to avoid this issue.

Example of the issue in Jelly: https://imgur.com/a/DT98ahR

Change-Id: Ifd906f5417cc0a090a2bf7d1f0dcdab26bf62dbb
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>